move into encoder_layouts.rs
